Created DeployPreview component
